services/evaluator/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
psycopg2-binary==2.9.9
